Refactoring Code - Move Api's from serializers.py file to views.py file - Add documentation in serializers and ViewSets - Remove unnecessary imports
